# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger UI for API testing",
        "title": "Wallet Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/aggregator/quotes": {
            "post": {
                "description": "并发调用所有可用 provider，返回最佳报价及候选列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aggregator"
                ],
                "summary": "获取聚合报价",
                "parameters": [
                    {
                        "description": "报价请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend.QuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend.QuoteResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/aggregator/swap/prepare": {
            "post": {
                "description": "根据报价 ID 和用户地址生成签名动作链路",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aggregator"
                ],
                "summary": "生成 swap 执行计划",
                "parameters": [
                    {
                        "description": "prepare 请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend.PrepareSwapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend.PrepareSwapResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/aggregator/swap/status": {
            "get": {
                "description": "根据 swapId 返回各步骤最新状态，并在需要时刷新链上信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aggregator"
                ],
                "summary": "查询 swap 状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Swap ID",
                        "name": "swapId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend.SwapStatusResponse"
                        }
                    },
                    "400": {
                        "description": "swapId is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/aggregator/swap/submit_tx_hash": {
            "post": {
                "description": "用户使用前端钱包（如 MetaMask）自行广播交易后，将 txHash 回传给后端。后端记录 swap step 的 txHash 并将状态置为 PENDING，后续可通过 GetSwapStatus/worker 跟踪链上结果。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aggregator"
                ],
                "summary": "提交交易哈希（前端钱包已广播）",
                "parameters": [
                    {
                        "description": "提交交易哈希请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend.SubmitTxHashRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend.SubmitTxHashResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/aggregator/tx/submitSigned": {
            "post": {
                "description": "接收客户端签名后的原始交易并广播，同时更新步骤状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Aggregator"
                ],
                "summary": "提交签名交易",
                "parameters": [
                    {
                        "description": "签名交易请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend.SubmitSignedTxRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend.SubmitSignedTxResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/authenticator/generate": {
            "post": {
                "description": "生成60秒周期的TOTP密钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "身份认证"
                ],
                "summary": "生成TOTP密钥",
                "parameters": [
                    {
                        "description": "生成TOTP请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GenerateTOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "生成成功",
                        "schema": {
                            "$ref": "#/definitions/models.GenerateTOTPResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/authenticator/verify": {
            "post": {
                "description": "验证60秒周期的TOTP验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "身份认证"
                ],
                "summary": "验证TOTP代码",
                "parameters": [
                    {
                        "description": "验证TOTP请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VerifyTOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证成功",
                        "schema": {
                            "$ref": "#/definitions/models.VerifyTOTPResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "backend.Action": {
            "type": "object",
            "properties": {
                "action_type": {
                    "$ref": "#/definitions/backend.ActionType"
                },
                "chain_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "signing_payload": {
                    "$ref": "#/definitions/backend.SigningPayload"
                }
            }
        },
        "backend.ActionType": {
            "type": "string",
            "enum": [
                "APPROVE",
                "SWAP",
                "BRIDGE",
                "WRAP",
                "UNWRAP"
            ],
            "x-enum-comments": {
                "ActionTypeApprove": "授权操作",
                "ActionTypeBridge": "跨链桥接",
                "ActionTypeSwap": "交换操作",
                "ActionTypeUnwrap": "解包装代币",
                "ActionTypeWrap": "包装原生代币"
            },
            "x-enum-descriptions": [
                "授权操作",
                "交换操作",
                "跨链桥接",
                "包装原生代币",
                "解包装代币"
            ],
            "x-enum-varnames": [
                "ActionTypeApprove",
                "ActionTypeSwap",
                "ActionTypeBridge",
                "ActionTypeWrap",
                "ActionTypeUnwrap"
            ]
        },
        "backend.ChainType": {
            "type": "string",
            "enum": [
                "EVM",
                "SOLANA"
            ],
            "x-enum-varnames": [
                "ChainTypeEVM",
                "ChainTypeSolana"
            ]
        },
        "backend.PrepareSwapRequest": {
            "type": "object",
            "required": [
                "best_quotes_index",
                "quote_id"
            ],
            "properties": {
                "best_quotes_index": {
                    "type": "integer"
                },
                "quote_id": {
                    "type": "string"
                }
            }
        },
        "backend.PrepareSwapResponse": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend.Action"
                    }
                },
                "swap_id": {
                    "type": "string"
                }
            }
        },
        "backend.Quote": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "type": "string"
                },
                "chain_type": {
                    "$ref": "#/definitions/backend.ChainType"
                },
                "fees": {
                    "type": "string"
                },
                "from_amount": {
                    "type": "string"
                },
                "from_token": {
                    "type": "string"
                },
                "gas_estimate": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "raw": {
                    "description": "Raw provider response",
                    "type": "string"
                },
                "router": {
                    "description": "EVM only: swap router",
                    "type": "string"
                },
                "spender": {
                    "description": "EVM only: approval spender",
                    "type": "string"
                },
                "to_amount": {
                    "type": "string"
                },
                "to_token": {
                    "type": "string"
                }
            }
        },
        "backend.QuoteRequest": {
            "type": "object",
            "required": [
                "amount",
                "from_chain_id",
                "from_token",
                "slippage_bps",
                "to_chain_id",
                "to_token"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "from_chain_id": {
                    "type": "string"
                },
                "from_token": {
                    "type": "string"
                },
                "slippage_bps": {
                    "type": "number",
                    "maximum": 10000,
                    "minimum": 0
                },
                "to_chain_id": {
                    "type": "string"
                },
                "to_token": {
                    "type": "string"
                },
                "user_address": {
                    "type": "string"
                },
                "wallet_uuid": {
                    "description": "Optional: wallet UUID for tracking",
                    "type": "string"
                }
            }
        },
        "backend.QuoteResponse": {
            "type": "object",
            "properties": {
                "best_quotes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend.Quote"
                    }
                },
                "expires_at": {
                    "type": "string"
                },
                "quote_id": {
                    "type": "string"
                },
                "user_address": {
                    "type": "string"
                },
                "wallet_uuid": {
                    "type": "string"
                }
            }
        },
        "backend.SigningPayload": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "gas": {
                    "type": "string"
                },
                "serialized_tx": {
                    "description": "Solana fields",
                    "type": "string"
                },
                "to": {
                    "description": "EVM fields",
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "backend.Step": {
            "type": "object",
            "properties": {
                "action_type": {
                    "$ref": "#/definitions/backend.ActionType"
                },
                "confirmed_at": {
                    "type": "string"
                },
                "expected_chain_id": {
                    "description": "用 string 存更通用",
                    "type": "string"
                },
                "expected_data_hash": {
                    "description": "0x...",
                    "type": "string"
                },
                "expected_to": {
                    "type": "string"
                },
                "expected_value": {
                    "description": "wei，十进制或 hex 统一一种",
                    "type": "string"
                },
                "fail_message": {
                    "type": "string"
                },
                "fail_reason_code": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "status": {
                    "description": "0=CREATED, 1=PENDING, 2=FAILED, 3=SUCCESS",
                    "type": "integer"
                },
                "step_index": {
                    "type": "integer"
                },
                "submitted_at": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                }
            }
        },
        "backend.SubmitSignedTxRequest": {
            "type": "object",
            "required": [
                "idempotency_key",
                "signed_tx",
                "step_index",
                "swap_id"
            ],
            "properties": {
                "idempotency_key": {
                    "type": "string"
                },
                "signed_tx": {
                    "type": "string"
                },
                "step_index": {
                    "type": "integer",
                    "minimum": 0
                },
                "swap_id": {
                    "type": "string"
                }
            }
        },
        "backend.SubmitSignedTxResponse": {
            "type": "object",
            "properties": {
                "tx_hash": {
                    "type": "string"
                }
            }
        },
        "backend.SubmitTxHashRequest": {
            "type": "object",
            "required": [
                "idempotency_key",
                "swap_id",
                "tx_hash"
            ],
            "properties": {
                "idempotency_key": {
                    "type": "string"
                },
                "step_index": {
                    "type": "integer",
                    "minimum": 0
                },
                "swap_id": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                }
            }
        },
        "backend.SubmitTxHashResponse": {
            "type": "object",
            "properties": {
                "tx_hash": {
                    "type": "string"
                }
            }
        },
        "backend.SwapStatusResponse": {
            "type": "object",
            "properties": {
                "fail_message": {
                    "type": "string"
                },
                "fail_reason_code": {
                    "type": "string"
                },
                "status": {
                    "description": "0=CREATED, 1=PENDING, 2=FAILED, 3=SUCCESS",
                    "type": "integer"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend.Step"
                    }
                },
                "swap_id": {
                    "type": "string"
                }
            }
        },
        "models.GenerateTOTPRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.GenerateTOTPResponse": {
            "type": "object",
            "properties": {
                "current_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "qr_code_image": {
                    "type": "string"
                },
                "qr_code_url": {
                    "type": "string"
                },
                "remaining_time": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "models.VerifyTOTPRequest": {
            "type": "object",
            "required": [
                "code",
                "secret"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "models.VerifyTOTPResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "remaining_time": {
                    "type": "integer"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    }
}